+++ src/main/java/org/yinwang/yin/Constants.java
@@ -32,6 +32,7 @@ public class Constants {
     public static final String DEF_KEYWORD = "define";
     public static final String ASSIGN_KEYWORD = "set!";
     public static final String RECORD_KEYWORD = "record";
+    public static final String FIELD_KEYWORD = "field";
     public static final String DECLARE_KEYWORD = "declare";
     public static final String UNION_KEYWORD = "U";