+++ src/test/java/org/yinwang/yin/TypeCheckerIntegrationTest.java
@@ -26,6 +26,7 @@ class TypeCheckerIntegrationTest {
         assertEquals("Int", typecheck("tests/function1.yin").toString());
         assertEquals("Int", typecheck("tests/recursion-direct.yin").toString());
         assertEquals("void", typecheck("tests/recursion-mutual.yin").toString());
+        assertEquals("Int", typecheck("tests/record-field-access.yin").toString());
     }
 
     @Test